mas cosas actividad 1

+++ skeleton/GravityForceGenerator.cpp
@@ -2,5 +2,5 @@
 
 void GravityForceGenerator::applyForce(Particle* p)
 {
-	p->setAcc({ 0, p->getMass() * g, 0 });
+	p->setAcc({ 0, (float)p->getMass() * g, 0 });
 }
